backend/requirements.txt
# Python Backend Requirements
# Install with: pip install -r requirements.txt

# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
PyMySQL==1.1.0
mysqlclient==2.2.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# Data Validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Environment Variables
python-dotenv==1.0.0

# Development
Flask-Migrate==4.0.5
